Move data getters to util file since they are not related to the page itself
